+++ tests/test_gsc_bookmarks.py
@@ -74,6 +74,8 @@ class GoogleSearchConsoleBookMarkTest(GoogleSearchConsoleBaseTest):
                         "image": second_sync_bookmark_date,
                         "web": second_sync_bookmark_date,
                         "video": second_sync_bookmark_date,
+                        "news": second_sync_bookmark_date,
+                        "googleNews": second_sync_bookmark_date,
                     }
                 },
                 "performance_report_page": {
@@ -81,6 +83,9 @@ class GoogleSearchConsoleBookMarkTest(GoogleSearchConsoleBaseTest):
                         "image": second_sync_bookmark_date,
                         "web": second_sync_bookmark_date,
                         "video": second_sync_bookmark_date,
+                        "discover": second_sync_bookmark_date,
+                        "news": second_sync_bookmark_date,
+                        "googleNews": second_sync_bookmark_date,
                     }
                 },
                 "performance_report_country": {
@@ -88,6 +93,9 @@ class GoogleSearchConsoleBookMarkTest(GoogleSearchConsoleBaseTest):
                         "image": second_sync_bookmark_date,
                         "web": second_sync_bookmark_date,
                         "video": second_sync_bookmark_date,
+                        "discover": second_sync_bookmark_date,
+                        "news": second_sync_bookmark_date,
+                        "googleNews": second_sync_bookmark_date,
                     }
                 },
                 "performance_report_query": {
@@ -95,6 +103,7 @@ class GoogleSearchConsoleBookMarkTest(GoogleSearchConsoleBaseTest):
                         "image": second_sync_bookmark_date,
                         "web": second_sync_bookmark_date,
                         "video": second_sync_bookmark_date,
+                        "news": second_sync_bookmark_date,
                     }
                 },
                 "performance_report_date": {
@@ -102,6 +111,9 @@ class GoogleSearchConsoleBookMarkTest(GoogleSearchConsoleBaseTest):
                         "image": second_sync_bookmark_date,
                         "web": second_sync_bookmark_date,
                         "video": second_sync_bookmark_date,
+                        "discover": second_sync_bookmark_date,
+                        "news": second_sync_bookmark_date,
+                        "googleNews": second_sync_bookmark_date,
                     }
                 },
                 "performance_report_custom": {
@@ -109,6 +121,9 @@ class GoogleSearchConsoleBookMarkTest(GoogleSearchConsoleBaseTest):
                         "image": second_sync_bookmark_date,
                         "web": second_sync_bookmark_date,
                         "video": second_sync_bookmark_date,
+                        "discover": second_sync_bookmark_date,
+                        "news": second_sync_bookmark_date,
+                        "googleNews": second_sync_bookmark_date,
                     }
                 },
             },